Consensus as points
Changed consensus species to be static points. The interpolated points will be a separate asset. So the idea is there are points for the table top (leaves), points in 3D for the consensus species (this), and points in-between (interpolatable). The only completed one at this point is the consensus species points (this commit).