Drawing branches
 - WIP stuff for drawing branches inside the sphere of points. DOESN'T WORK RIGHT - scaling is all wonky.

THIS IS THE LAST CHECKIN BEFORE REWORKING POSITIONING AND SCALING. After this, all datasets will be centered at 0,0,0 and provide (if applicable) a parent dataset and point to center the data at. This will achieve the "nested" model we've been talking about.

The Makefile contains changes for the start of this, including args for using a JSON file instead of a CSV file for input dataset "commands".